Split methods and unit tests related to database introducing MySQL connections
