+++ backend/interface/database_api/Cargo.toml
@@ -19,7 +19,7 @@ backend_logging = { path = "../../backend_logging" }
 backend_macros = { path = "../../backend_macros" }
 database_types = { path = "../../database_types" }
 chrono = "0.4.31"
-diesel = { version = "2.2.6", features = ["chrono", "postgres"] }
+diesel = { version = "2.2.6", features = ["chrono", "postgres", "mysql"] }
 flexi_logger = { version = "0.29.8", features = [
     "async",
     "colors",
