+++ backend/interface/database_api/src/api/member.rs
@@ -1,5 +1,5 @@
-use crate::api::models::Member;
-use crate::database::{bind_column_value, get_connection};
+use crate::database::bind_column_value;
+use crate::{api::models::Member, connection::get_connection};
 use database_types::connection_description::ConnectionDescription;
 use diesel::{RunQueryDsl, SelectableHelper, query_dsl::methods::SelectDsl, sql_types::Integer};
 use log::{error, info, warn};
