+++ backend/interface/database_api/src/lib.rs
@@ -1,5 +1,10 @@
 pub mod api;
 
+mod column_type_info;
+mod connection;
 mod database;
 mod frb_generated;
 mod schema;
+
+#[cfg(test)]
+mod test_database_common;
